Change integer data type from int64_t to int
- Modified data_s struct to use int instead of int64_t
- Updated ar_data_create_integer and ar_data_get_integer signatures
- Updated all format specifiers from %lld to %d
- Fixed variable declarations in tests and other modules
- Updated documentation to reflect the new data type

+++ modules/agerun_agency.c
@@ -180,8 +180,8 @@ bool ar_agency_load_agents(void) {
                     
                     data_t *value = NULL;
                     if (strcmp(type, "int") == 0) {
-                        int64_t int_value;
-                        if (fscanf(fp, "%lld", &int_value) != 1) {
+                        int int_value;
+                        if (fscanf(fp, "%d", &int_value) != 1) {
                             printf("Error: Could not read int value\n");
                             break;
                         }
